Split update 2.1.8 into its constituent modules and enabled interactive updates for CI.
