+++ .travis.yml
@@ -81,6 +81,8 @@ install:
   - cd docroot
   # Execute database updates, if there are any.
   - drupal update:execute
+  # Execute manual updates.
+  - drupal update:lightning --no-interaction
 
 before_script:
   - drush runserver --default-server=builtin 8080 &>/dev/null &
